Patch 6.1.404 (extra) Problem:    Various small problems. Solution:   Fix comments.  Various small additions, changes in indent, removal 	    of unused items and fixes. Files:	    Makefile, README.txt, runtime/menu.vim, runtime/vimrc_example.vim, 	    src/INSTALL, src/INSTALLole.txt, src/Make_bc5.mak, 	    src/Make_cyg.mak, src/Make_ming.mak, src/Makefile, 	    src/config.h.in, src/edit.c, src/eval.c, src/ex_cmds2.c, 	    src/ex_docmd.c, src/ex_getln.c, src/fileio.c, src/getchar.c, 	    src/gui.c, src/gui_gtk.c, src/gui_photon.c, src/if_cscope.c, 	    src/if_python.c, src/keymap.h, src/mark.c, src/mbyte.c, 	    src/message.c, src/misc1.c, src/misc2.c, src/normal.c, 	    src/option.c, src/os_os2_cfg.h, src/os_win32.c, 	    src/proto/getchar.pro, src/proto/message.pro, 	    src/proto/regexp.pro, src/screen.c, src/structs.h, src/syntax.c, 	    src/term.c, src/testdir/test15.in, src/testdir/test15.ok, 	    src/vim.rc, src/xxd/Make_cyg.mak, src/xxd/Makefile

+++ src/if_cscope.c
@@ -124,7 +124,7 @@ do_scscope(eap)
 {
     postponed_split = -1;
     do_cscope(eap);
-} /* do_scscope */
+}
 
 /*
  * PUBLIC: do_cstag
@@ -223,7 +223,7 @@ cs_fgets(buf, size)
     if ((p = cs_manage_matches(NULL, NULL, -1, Get)) == NULL)
 	return TRUE;
 
-    if (strlen(p) > size)
+    if ((int)strlen(p) > size)
     {
 	strncpy((char *)buf, p, size - 1);
 	buf[size] = '\0';
